fix(preprocessing): do not submit NaN or Infinity (#4359)
This PR changes the pipelines not to emit `NaN` and `Infinity` when
serializing JSON. `NaN` and `Infinity` are not part of JSON according to
the standard and are not accepted by the backend. If the user submits
"NaN" for a float field, it will be set to `null`. If they submit
`Infinity` or `-Infinity`, an error is thrown.