chore(website): rename `submissionId` to `id` for input (#3711)
<!-- Mention the issue that this PR resolves. Delete if there is no
corresponding issue -->
resolves #2404

Asking users to supply a metadata file with a "submissionId" for each
sequence is confusing. The ID here is actually likely to be the user's
own sample ID, as used in their FASTA file. "SubmissionID" implies an
association specifically with the Loculus submission - maybe an ID for
submission that you get from Loculus.

Here we rename this input field to a bare `id`, and also adjust the
templates accordingly. We have back-compatibility that also supports
`submissionId` (we can remove that after a while:
https://github.com/loculus-project/loculus/issues/4296 it's just in case
someone is preparing data right as we roll this out).

This does not rename the field downstream of submission or in our
internal variables. After submission this name does make more sense, it
distinguishes the name that Loculus assigned (the accession) from the
one the user used at submission (the `submissionId`).

This PR adjusts ingest to submit using `id`, but when ingest is
calculating hashes it renames this `id` back to `submissionId` for
compatibility with the old hashes.

# Testing

This is well covered by a large set of tests, and ingest by the preview
instance, additionally the following manual tests have been done:
- [x]  manually submitting example mpox data

I ~propose~ have also done the following 

- [x] putting this branch onto pathoplexus staging and observing that we
don't get version bumps, either on CCHF (segmented) or on non-segmented
organisms
    - I checked that ingest appeared to be running with no errors
- I checked that the number of sequences was the same as prod for all
organisms using the admin dashboard (meaning no new versions -[I also
checked for errors in preprocessing])

🚀 Preview: https://submissionid2id.loculus.org

---------

Co-authored-by: Loculus bot <bot@loculus.org>
Co-authored-by: Cornelius Roemer <cornelius.roemer@gmail.com>