feat(website): add dynamic banner message (#4356)
We have often encountered times when we have wanted to briefly have a
message on the website e.g. `We are currently rolling out a new release,
there may be brief disruptions to service`. One wants to be able to do
that without needing to change the underlying code. Here we achieve this
by allowing specifying a URL for the website to poll. The website caches
response.

Manual testing:

- I checked that the response time did not appear different to `main` in
the preview, suggesting caching is working correctly [then I also
checked by adding logging in a local version]

🚀 Preview: https://codex-add-banner-message.loculus.org

---------

Co-authored-by: Loculus bot <bot@loculus.org>
Co-authored-by: Chaoran Chen <mail@chaoran-chen.de>