feat(deployment): Make it possible to use local IP instead of `localhost` (#4460)
based on #4448, #4450, #4453

In #4344, #4298 we ran into issues with the integration tests running in
CI. After some research we found that the issues were caused by the
backend giving pre-signed URLs to the prepro pipeline, that only work
from outside the cluster.

This PR fixes that, by making it possible to use the local IP of the
host instead of the literal `localhost`. The IP will work inside and
outside of the cluster.

- Introduces new `localHost` value in Helm Chart to override `localhost`
literal with something else.
- Introduces `./deploy.py helm --use-localhost-ip` flag to use the IP
instead of the `localhost` literal.
- Introduces `./deploy.py cluster --bind-all` flag to bind to all
interfaces to reach bound ports via the IP.
- Uses the new flags in integration tests CI.

### PR Checklist
- [x] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- ~~Any manual testing that has been done is documented (i.e. what
exactly was tested?)~~

🚀 Preview: Add `preview` label to enable