fix(website): make rsuite date picker work acceptably with keyboard control (#4477)
This PR reverts back to using rsuite's date picker but with key changes
that fix various bugs (I'm not categorically opposed to using a
different date picker, or as in this case rolling our own but there's a
lot of edge cases that need handling and the current version doesn't
work on mobile, for example - so it's best in my view to go back to the
tried and tested way with key fixes and develop the other without
pressure).

The PR targets the "revert-custom-date" branch which has the custom date
picker reverted to make the diff here easier to review. The custom date
picker came with test improvements that are worth keeping in either
case.

The switch to `yyyy-MM-dd` is in separate PR #4480 which targets this
branch.

I switch back to uncontrolled mode as controlled mode plays badly with
keyboard input, possibly because that mode is not well tested,
especially when partial/early dates are visible in the field, such as
`11-01-0019` which is inevitable when using the keyboard.

Lots of keyboard entry bugs (#4425, #4084, unreported inability to type
for timestamp field) were introduced when we switched to controlled mode
in https://github.com/loculus-project/loculus/pull/3895. The switch to
controlled mode was presumably made to allow clearing the date field
using the search facet's x:
<img width="265" alt="image"
src="https://github.com/user-attachments/assets/d6a0ab62-5925-4b8d-8685-d962365ef6ff"
/>

This PR uses conditional rerendering of the DatePicker to control _only_
clearing when we want it: when the fieldValue changes from upstream.

I also rewrite the JSDate <-> IsoDate logic in the DateField component
to allow round tripping. This was essential to fix #4425 while using
controlled date picker. It's no longer essential for the uncontrolled
date picker - but nice to have.

### Manual testing

- Verify dates can be typed in properly in the date picker in both date
range and timestamp modes (both didn't work before)
- Clearing works: via cross on date picker, and facet above the search
ui
- Input doesn't lose focus on backspace
- Value is correctly set from URL params and vice versa