feat(int-tests): speed up read-only dependent tests by using pre-existing sequences (#4523)
When doing Playwright debugging locally, it was annoying that for
dependent tests, there would be new sequence creation every time one
runs a test. This took 15+ seconds.

This PR now checks for existence of a sequence and skips submission of
sequence if there already is a sequence in a group. This takes only ~2
seconds allowing faster iteration on dependent tests.

🚀 Preview: Add `preview` label to enable