test(integration-tests): fail on console warnings/errors (#4469)
Adds a fixture that keeps track of console warnings/errors in the
browser during the integration tests and fail if there are any.

There are certain messages we still ignore:
https://github.com/loculus-project/loculus/issues/4536 `Form submission
canceled because the form is not connected`
https://github.com/loculus-project/loculus/issues/4535 all
errors/warnings in Firefox

Nevertheless, this is a big improvement over the former state where were
ignoring everything!

🚀 Preview: Add `preview` label to enable