fix(backend): Fix file publishing for pipeline update race condition (#4581)
resolves #4580

This PR changes the file publishing code to not only publish files for
the current pipeline version, but also for all larger pipeline versions
(even if they are not the current one yet). This means that if the
current pipeline version is switched later on, the file will already be
public.

The PR includes a test - I checked that the test failed without the
change applied.

**manual testing** - If I'm not mistaken, this requires quite a bit of
luck to get the timing right, so it's not trivial to test manually. IMO
the automated test makes me very confident in the fix.

### PR Checklist
- ~~All necessary documentation has been adapted.~~
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: Add `preview` label to enable

---------

Co-authored-by: Chaoran Chen <mail@chaoran-chen.de>