chore(deps): bump kotlinx-datetime in backend (#4585)
changelog:
https://github.com/Kotlin/kotlinx-datetime/blob/master/CHANGELOG.md#070
Migration info:
https://github.com/Kotlin/kotlinx-datetime#deprecation-of-instant

There was a breaking change, where the `kotlinx` library deprecated
their own `Clock` and `Instant` in favor of the std lib implementations.
Using the std lib implementations however still requires
[`ExperimentalTime`](https://kotlinlang.org/api/core/kotlin-stdlib/kotlin.time/-experimental-time/)
- I think. That's why I added the compiler opt in option.
Without it, all places that use time now raise this comile error:

e: file:///...SeqSetCitationsDatabaseService.kt:314:32 This declaration
needs opt-in.
Its usage must be marked with '@kotlin.time.ExperimentalTime' or
'@OptIn(kotlin.time.ExperimentalTime::class)'

Also, I tried to not use the `-0.6.x-compat` version of the `0.7.0`, but
I got a bunch of runtime errors. However, this compat version will not
be supplied for `0.8.0`, so in the meantime we'd need to see that we
migrate away from it. We need to find exactly the libraries that might
be causing issues.

I think, the issue is `exposed-kotlin-datetime`:
https://github.com/JetBrains/Exposed/blob/89e0be3139a36df8106a6c20b28960cc6cd44408/gradle/libs.versions.toml#L40

Might be good for someone else to have a look too if I understood the
Changelog correctly.

If I got it right, I'd recommend we create an issue to check back on
bumping away from the compat version.

### Screenshot

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [ ] The implemented feature is covered by appropriate, automated
tests.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: Add `preview` label to enable

---------

Co-authored-by: Chaoran Chen <mail@chaoran-chen.de>