feat(deposition): fix model validation error (#4655)
resolves https://github.com/loculus-project/loculus/issues/4654

The refactor added a requirement on the get-group-info response, it
expected a json but the response was parsed into a python dict causing
failure. This fixes the error. I also tested the full submission on this
preview to ensure there were no other hidden errors.

This additionally adds support for the non-broker case by not filling
the author or address manifest field in these cases (ENA does not allow
these fields for non-brokers)

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [ ] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)
tested on the preview by uploading sequences, running cronjob and adding
results to a branch in the ena-submission repo (linked in the config for
testing), checked submission to dev succeeded

🚀 Preview: https://fix-group-validation.loculus.org