feat(deposition): send GCA accession to backend (#4695)
resolves https://github.com/loculus-project/loculus/issues/2945

### Summary
As we finally added the GCA accession to PPX:
https://github.com/pathoplexus/pathoplexus/pull/201 we can now also
upload it with the other submission results.

As https://github.com/loculus-project/loculus/pull/4688 is submitted
this is safe to do.

This PR resets the state of submission to SUBMITTED_ALL so that the
external metadata results will be resent and thus updated in Loculus

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [ ] The implemented feature is covered by appropriate, automated
tests.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: Add `preview` label to enable