feat(deposition): don't revise assembly if no change, remove all mypy type errors (#4652)
1. When revising a sequence check if there is a difference in the
assembly, by looking for differences in the
`DEFAULT_EMBL_PROPERTY_FIELDS` or the actual sequence - if there is no
change do not revise the sequence just copy over the results from the
previous version. If there has been a change revise the sequence as
normal
1. Adds an integration test that the assembly is not revised if there is
no change

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [ ] The implemented feature is covered by appropriate, automated
tests.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://dont-revise-assembly-if-n.loculus.org

---------

Co-authored-by: Cornelius Roemer <cornelius.roemer@gmail.com>