feat!(kubernetes): [CFG] add support for multiple reference genomes (#4692)
resolves #4691 

Breaking change in the config:

The reference genomes became a dictionary `[suborganismName,
referenceGenome]`. Before:
```yaml
    referenceGenomes:
        nucleotideSequences:
          - name: 'main'
            sequence: 'NNN'
        genes: []
```

After:
```yaml
    referenceGenomes:
      singleReference:
        nucleotideSequences:
          - name: 'main'
            sequence: 'NNN'
        genes: []
```

The key must be `singleReference` if there only one suborganism. (Maybe
we'll bind logic to that specific name and it's easier to drop that
requirement later compared to adding it)


This is supposed to be minimally invasive and change as few place in the
code as possible. For now this simply assumes that each organism has
exactly one reference genome configured (handling multiple ones in the
code will be the next step).

### PR Checklist
- [x] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://4691-multi-pathogen-suppo.loculus.org