feat(kubernetes): merge multi pathogen reference genomes for LAPIS (#4753)
resolves #4710

Without #4721, the ingest won't work, but from the SwaggerUI the LAPIS
instance looks good.

This takes all reference genomes and merges them into a single one in
the multi pathogen case. I'm planning to also put the same merged
reference genomes into the backend in the next step, because that's the
schema that it has to validate so that SILO will accept the data.

### Screenshot

### PR Checklist
~~- [ ] All necessary documentation has been adapted.~~
  - is there anything?
~~- [ ] The implemented feature is covered by appropriate, automated
tests.~~
  - I don't think it's possible to test this?
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)
  - Manually clicking around on the preview. Nothing specific.

🚀 Preview: https://4710-multi-pathogen-suppo.loculus.org