test(prepro): add unit tests that use nextclade (#4696)
resolves #

### Screenshot

As we will be making more changes to prepro (with the annotations and
the multi-organism case) I thought it would be good to add some tests
for our current alignment functionalities I also noticed some slightly
inconsistent behavior that we can clean up a bit.

For single configs:
- [x] add test for insertion, deletion and mutation
- [x] add test for failed alignment

For multi-segment configs:
- [x] add test for insertion, deletion and mutation
- [x] add test for failed alignment
- [x] add test for all segments failed
- [x] add test for one failed at least one other succeeded
- [x] add test for not all segments uploaded


### PR Checklist
- [x] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: Add `preview` label to enable

---------

Co-authored-by: Cornelius Roemer <cornelius.roemer@gmail.com>