feat(backend): multipart file upload (#4378)
This implements the backend for #3993. It depends on #4380.

To use multipart uploads, the workflow is:

1. Call `POST
/files/request-multipart-upload?numberFiles=<...>&numberParts=<...>` –
this returns a list of File IDs and, for each File ID, a list of
presigned URLs to upload the parts. The URLs are ordered, the first link
is for the first part, etc.
2. Use `PUT` to upload the parts. Each response will contain an `etag`
in the header. Not all URLs/parts need to be used, but $\forall i:\
\text{if } \mathit{part}_{i+1} \text{ is used} \Rightarrow
\mathit{part}_i \text{ is used}$.
3. Call `POST /files/complete-multipart-upload`, sending `[{fileId: ...,
etags: [...]}, ...` in the body
4.Submitting, revising, and submitting processed data expect that
uploads have been completed.

### PR Checklist
- [x] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)
- I manually uploaded 2 files with each 2 parts, submitted and approved
them, and downloaded them from the website. I verified that the files
were correct.

🚀 Preview: https://multipart-upload.loculus.org