feat(backend): only add the merged reference genomes to the backend config (#4770)
Instead of the full non-merged reference genomes, only supply the merged
one to the backend. That should give as a working preview with
entero-viruses fastest.
The validation of the preprocessing results will simply work with that.
The initial submission will still be in the multi segmented style, but
we can tackle that later.

Apparently requires https://github.com/loculus-project/loculus/pull/4767
to have working ingest for the EVs.

This is a partial revert of #4692:
- The values.yaml config changes stay (extra level of hierarchy for each
suborganism)
- The rename in backend of `referenceGenomes` -> `referenceGenome` stays
but the backend config goes back to what it was before, flat.
- Website changes stay

### PR Checklist
- [x] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)
  - the organisms on the preview still seem to work as before.

🚀 Preview: https://onlymergedrefgenomeforbac.loculus.org