feat(ingest): never remove segment field from metadata and filter out nextclade sort matches that are not supported (#4767)
resolves #

Main is failing with the error 
```
 File "/package/scripts/heuristic_group_segments.py", line 145, in main
    segment = values["segment"]
              ~~~~~~^^^^^^^^^^^
KeyError: 'segment'
```
`segment` is a keyword in the heuristic grouping script. However, in the
enteroviruses config we renamed `segment` to genotype. This rename meant
that `segment` was nolonger a field in the metadata - leading the
heuristic grouping script to break. This PR should ensure that even if
we rename `segment` the `segment` field should stay in the metadata as
it is a required field for grouping.

This PR also covers the case where nextclade sort assigns a segment to a
sequence which is not a valid segment - in this case we also drop these
matches. This is important as the backend (for now) does not allow
submission of unknown segments and will error - also ingest has been
written with this state in mind and assumes we only assign known
segments.

This was fixed in @alejandra-gonzalezsanchez's PR:
https://github.com/loculus-project/loculus/pull/4570

### Screenshot

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [ ] The implemented feature is covered by appropriate, automated
tests.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://small-ingest-fixes.loculus.org

---------

Co-authored-by: Cornelius Roemer <cornelius.roemer@gmail.com>