feat(backend): omit nulls in insertions and sequences and do not require prepro to upload null sequences (#4815)
resolves https://github.com/loculus-project/loculus/issues/4814,
resolves https://github.com/loculus-project/loculus/issues/4763

### Screenshot

Similar to in https://github.com/loculus-project/loculus/pull/3232 now
that we have multi-pathogens we also want to facilitate the easy
addition of new "segments" (i.e. suborganisms) and their genes.

We can do this again in an efficient way where we do not store nulls in
the backend but impute the `null` entries while streaming - this also
means that we can drop the requirement on prepro to add nulls for all
"missing" segments and genes - decreasing the amount of data we have to
stream.

This PR omits `null` entries and empty lists for nucleotide and amino
acid sequences as well as insertions.

### PR Checklist
- [x] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented: confirmed
that CCHF preview is same state as main as prepro no longer returns
`null` when a segment is missing this is a way to confirm that LAPIS is
still receiving the correct values.

🚀 Preview: https://sequence-hydration.loculus.org