feat(backend): Delete outdated preprocessing data (#4584)
resolves #3240 

In the task that periodically checks if the prepro pipeline version
needs to be bumped, we also do the cleaning now.
For every organism where the version has been bumped, we remove outdated
data from `sequence_entries_preprocessed_data`

The query to delete the data uses a subquery, because in the table, we
don't know which row belongs to which organism. We cannot do that in
Exposed directly, so it's done in raw SQL. IMO it would be better if we
had an `organism` column directly in the table - this would also help us
elsewhere (should we make an issue for that?)

**Backend unit test**: I've added a test that adds data for 3 processing
versions and then checks that only data for the last two is present. The
test also includes data for 2 organisms, and checks that only data for
one organism is deleted.

**manual Testing**: I created a persistent preview, and added a third
prepro pipeline, and checked in the DB if v1 data was deleted (it was)
but that it was still present for other organisms (it was). I also
checked the website, spot checked some organism entries that things look
unaffected (looked good).