feat(website): multi pathogen support: sequence details page: only show relevant mutation references (#4773)
resolves #4784

This PR contains:
* squashed versions of #4767 and #4770 because they are required to get
a working preview (they should be dropped before merging)
* compute the suborganism from the respective LAPIS field
* Still to do: How do we configure which field that should be? Add a
field to `defaultOrganisms.[organism].schema.website`?
* making the `ReferenceGenomesSequenceNames` a dictionary `{
[suborganism]: <the object with the names> } and propagate this through
the website (a lot of noise...)
* read the first value of that dictionary almost everywhere
* This is to keep the website in a working state until it's properly
implemented everywhere. The single pathogen case will stay untouched. In
the multi pathogen case, some things that are shown on the website might
be wrong until follow up PR "fix" it by properly implementing it.
* in the mutations section of the details page, only show the references
of the respective sequence entry:
  * `Mutations called relative to the -->this here<-- reference`


### Screenshot
<img width="949" height="413" alt="image"
src="https://github.com/user-attachments/assets/3ee3be36-1451-49e9-9acd-68250f169c32"
/>

### PR Checklist
~~- [ ] All necessary documentation has been adapted.~~
- [ ] The implemented feature is covered by appropriate, automated
tests.
- I did not find any unit test for the part where the data table is
shown to check that it shows the correct references there.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)
- For enterovirus entries with different genotypes, it only shows the
reference of the suborganism (the one that is in the reference genome)
  - For other organisms, it still shows the correct references, too.

🚀 Preview: https://4772-multi-pathogen-suppo.loculus.org