feat!(config): support version lists in preprocessing pipeline config for better developer experience (#5083)
resolves #4703

### Config changes (non-breaking)

Support version as list in preprocessing part of config

### Description

Bumping prepro pipeline versions has been a bit of a pain due to having
to copy/paste/delete or comment/uncomment entries - in particular also
when nothing changed but the version (which is the most common case).

This PR makes bumping prepro versions easier by allowing version to be a
list, which acts as syntactic sugar for a copy for each entry in the
list with version substituted by each item's value.

To catch mistakes, we validate that each version appears at most once
and throw an error if that's not the case.

This change is fully backwards compatible. What's *breaking* is that we
can now take advantage of this new feature!

Example usage:
```yaml
preprocessing:
  - version: [1, 2, 3]  # Creates 3 separate deployments
    image: ghcr.io/loculus-project/preprocessing-nextclade
  - version: 4          # Creates 1 additional deployment
```

### Screenshot

<img width="1504" height="532" alt="image"
src="https://github.com/user-attachments/assets/6802daab-454c-4076-af90-33488dd72d68"
/>

### Manual testing

I tested that the validation for duplicates works.