fix(integration-test): Extend timeout to get options from LAPIS in search pages from 3s to 15s (#5121)
This _could_ be the cause of
https://github.com/loculus-project/loculus/issues/5096 (i.e. maybe LAPIS
is slow in some circumstances). It might well not be the problem at the
moment - but in any case it makes sense to allow more than 3 seconds in
case.

[Having tested this locally it definitely doesn't resolve all such
issues, but I think it makes clearer than the failure is not due to such
slowness]

🚀 Preview: Add `preview` label to enable