feat(backend): clean up auxTable entries that are older than 24 hours (#5189)
resolves https://github.com/loculus-project/loculus/issues/5135

If the backend crashes after uploading data to the aux table the data
will get stuck indefinitely as we have no way of cleaning it up. We
prevented this case indirectly by adding a Lock but as this is now
removed (which is good as we don't want submissions to be blocked for
all users if the backend crashed during submission from one user) we
should add a method for cleaning up the aux table.

### Screenshot

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [ ] The implemented feature is covered by appropriate, automated
tests.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: Add `preview` label to enable

---------

Co-authored-by: Theo Sanderson <theo@sndrsn.co.uk>