fix(ingest): move defaults to config (#5263)
resolves Missing sequences on main

### Screenshot
Ingest sets default config values from the default.yaml config - these
values override any defaults set inside a Config object used in a
script. If a default value is not set in the default config it becomes a
default empty string and will not be parsed correctly.

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [ ] The implemented feature is covered by appropriate, automated
tests.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://fix-ingest-timeout.loculus.org