fix(backend): Check first version of unversioned accession in seqset creation, not latest
When adding an unversioned accession to a seqset, we used to check if the latest version was released. This is a bug as seqset creation should not be dependent on the latest version being released.

An unversioned sequence should be allowed to be added as long as version one is released.

This should fix the flaky integration test #5113

The test revelaed a real bug - the fact it was flaky was due to the bug only being revealed when the sequence was in the process of revision.