fix(backend): check first version of unversioned accession in seqset creation, not latest (#5268)
When adding an unversioned accession to a seqset, we used to check if
the latest version was released. This is a bug as seqset creation should
not be dependent on the latest version being released.

An unversioned sequence should be allowed to be added as long as version
one is released.

This should fix the flaky integration test #5113

The test revealed a real bug - the fact it was flaky was due to the bug
only being revealed when the sequence was in the process of revision.

resolves #5113

The main code change is:
https://github.com/loculus-project/loculus/pull/5268/files#diff-09197b7be14ff0e50c30b0d4ca110ec76be66b5826720fd596a87498547293f2R471-R480

The rest is slight improvements to error messages in related code.

### Manual testing

- [x] Create seqset with versioned and unversioned accessions
- [x] Create seqset with unversioned accession where there's a v2
waiting for approval