feat(database): optimize external_metadata_view and sequence_entries_view (#5260)
resolves #5257 

The way the current pipeline version was fetched was too complicated.

On GenSpectrum we got the `select count(*) from sequence_entries_view`
from ~28 seconds to ~3 seconds;

co-written with @chaoran-chen and actually @theosanderson found the same
changes in his PR!
https://github.com/loculus-project/loculus/pull/5258/files

### Testing

I tested that the counts on the views are the same before and after.

Beyond that I expect the automated tests to be enough.

### PR Checklist
- ~~All necessary documentation has been adapted.~~
- ~~The implemented feature is covered by appropriate, automated
tests.~~
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://optimize-views.loculus.org

---------

Co-authored-by: GitHub Action <action@github.com>
Co-authored-by: Anna (Anya) Parker <50943381+anna-parker@users.noreply.github.com>