feat(deposition): Distinguish between complete and incomplete (raw read) ENA metadata in notifications
We currently triage all submissions that have any ENA specific metadata together.

It's better if we treat them differently depending on if _all_ vs _some_ of the ENA specific fields are present.

If all are present, this usually means users want us to link to their raw reads. If not all are present, this warrants further investigation.