feat(deposition): only block revision if formatted authors field changes rather than unformatted (#5298)
The ENA deposition pipeline blocks revisions if a metadata field changes
that is in the manifest. This is because we cannot revise manifest
fields through the API and need to notify ENA manually through email.

This PR updates the code to check if the formatted authors' list changes
(what we actually send to ENA) as opposed to the unformatted authors'
list.

See https://loculus.slack.com/archives/C05G172HL6L/p1761121808948159

Linked to https://github.com/loculus-project/loculus/issues/5296

🚀 Preview: Add `preview` label to enable