feat(backend): add request duration logging for streaming requests (#5347)
In https://github.com/loculus-project/loculus/pull/5242 we added
duration logging for requests, but for streamed requests the "200"
response is sent immediately, but the actual data can take quite a bit
longer.

This PR adds timing output for streamed requests. I'm also passing the
name of the endpoint to the function, so it's easy to spot when among
the long running requests there is one that usually shouldn't be there.

(I'm suspecting that `extract-unprocessed-data` is sometimes taking very
long to respond, but the logs are not definitive yet)

**manual testing** - I tested this in GenSpectrum and it's running fine
and I can see the numbers.

### PR Checklist
- ~~All necessary documentation has been adapted.~~
- ~~The implemented feature is covered by appropriate, automated
tests.~~
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://backend-time-streamed.loculus.org

---------

Co-authored-by: Fabian Engelniederhammer <92720311+fengelniederhammer@users.noreply.github.com>