feat(prepro): provide request ID in backend requests to allow tracing requests (#5346)
When a prepro Pod crashes with a timeout, we don't know what happened to
the request in the backend, or we can only correlate it by looking at
the time the Pod crashed.

The backend already reads the `x-request-id` header and uses the request
ID in it, if given.

This PR sets a UUID as a request ID in the requests sent to the backend.

**Manual testing** - I deployed this to GenSpectrum staging and looked
in loki and found a request successfully correlated.