feat(silo_import): use ruff and mypy, move to new folder, make version an int (#5377)
Main features of this update PR are:
1. Apply ruff and mypy to improve typing (also add to github actions)
2. Change version from type string to int
3. Move silo_import to new folder
4. Add more info to the README and use constants more consistently 
5. Logic change: if decompression fails keep current (= last successful)
etag (do not set to "0") -> no change in behavior, will still be retried
however the current_etag will be the etag of the currently used data in
SILO
6. Logic change: if hard refresh fails do not set the last hard refresh
time to the time of the failure (i.e. ensure the hard refresh is retried
until it succeeds)

### PR Checklist
- [x] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?) I just checked that the preview was healthy and we
only reprocess when there is a data change

🚀 Preview: https://pyimport-anya.loculus.org

---------

Co-authored-by: Theo Sanderson <theo@sndrsn.co.uk>